% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/covUitls.R
\name{compareCov}
\alias{compareCov}
\title{This is a utility function to compare two covariance matrices}
\usage{
compareCov(cov1, cov2, labels, corr = FALSE)
}
\arguments{
\item{cov1}{covariance matrix using the first method}

\item{cov2}{covariance matrix using the second method}

\item{labels}{strings indicating the type of methods used for comparison}

\item{corr}{flag indicating if the supplied matrices are of type covariance
or correlation}
}
\description{
This is a utility function to compare two covariance matrices
}
\details{
This method takes in two different covariance/correlation matrices computed
using two different methods and visullay compares them using the ellipse plot.
It produces a matrix with ellipses drawn in the upper triangle. The ellipse
is drawn to be a contour of a standard bivariate normal with correlation given
by the correlation of the two assets. One ellipse is drawn for each covariance
matrix.
}
\author{
Rohit Arora
}

