% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCoux.R
\name{.biweight}
\alias{.biweight}
\title{Biweight Rho function}
\usage{
.biweight(x, p)
}
\arguments{
\item{x}{the value}

\item{p}{number of columns}
}
\description{
Biweight Rho function
}
\details{
Calculate the output of a biweight rho function given the value and the number of 
variables
}
