% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.c.ell}
\alias{.c.ell}
\title{Equation 1.2}
\usage{
.c.ell(lambda, gamma)
}
\arguments{
\item{lambda}{sample eigenvalue}

\item{gamma}{fitted value of variables/observations}
}
\description{
Equation 1.2
}
\details{
Equation 1.2 as described in (Donoho, Gavish, and Johnstone, 2013)
}
