% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.detectSpikes}
\alias{.detectSpikes}
\title{Detect spikes in a covariance model}
\usage{
.detectSpikes(lambdas, numObs, alpha = 0.05)
}
\arguments{
\item{lambdas}{eigenvalues to which the distribution is fitted}

\item{numObs}{number of data observations}

\item{alpha}{significance level for the test}
}
\description{
Detect spikes in a covariance model
}
\details{
This method calculates the number of spikes in a spiked covariance model
by iteratively comparing a standardized eigenvalue to the qunatile of Tracy-Wisdom
distribution for given significance level. The procedure is described in 
(Kritchman and Nadler, 2009).
}
