% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.ell.lambda}
\alias{.ell.lambda}
\title{Equation 4.4}
\usage{
.ell.lambda(lambda, gamma)
}
\arguments{
\item{lambda}{sample eigenvalue}

\item{gamma}{fitted value of variables/observations}
}
\description{
Equation 4.4
}
\details{
Equation 4.4 as described in (Donoho, Gavish, and Johnstone, 2013)
}
