% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCoux.R
\name{.gamma_cp}
\alias{.gamma_cp}
\title{Calculation of constant required for consistency}
\usage{
.gamma_cp(p, c)
}
\arguments{
\item{p}{number of variables}

\item{c}{square root of the chi-square cutoff}
}
\description{
Calculation of constant required for consistency
}
\details{
The biweight rho function needs a constant whose value is determined based on
the consistency equation. Solving the integral and assuming multivariate
normality we can get an analytical expression for the constant.
}
