% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.getMPfit}
\alias{.getMPfit}
\title{Fitting an MP distribution to the data}
\usage{
.getMPfit(lambdas, gamma, numOfSpikes, numObs, method)
}
\arguments{
\item{lambdas}{eigenvalues of the sample covariance matrix}

\item{gamma}{ratio of varibales/observations}

\item{numOfSpikes}{number of spikes in the spike covariance model}

\item{numObs}{number of data observations}

\item{method}{KNTest/median fitting}
}
\description{
Fitting an MP distribution to the data
}
\details{
This method takes in the eigenvalues of the sample covariance matrix and fits 
a Marchenko Pastur distribution. The eigenvalues are assumed to have unit variance
}
