% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCoux.R
\name{.givens.rotation}
\alias{.givens.rotation}
\title{Givens rotation matrix}
\usage{
.givens.rotation(i, j, theta, d)
}
\arguments{
\item{i}{row number}

\item{j}{column number}

\item{theta}{angle in the range [-pi/2, pi/2]}

\item{d}{number of columns}
}
\description{
Givens rotation matrix
}
\details{
Givens rotation matrix is an orthogonal matrix with certain structure.
}
