% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentSwitchingDCC.R
\name{.helper.loglik}
\alias{.helper.loglik}
\title{Calculate the log-liklihood using Expectation Maximization}
\usage{
.helper.loglik(taus, theta1, theta2, returns, Sigma, numRegimes, Q, Q.bar)
}
\arguments{
\item{taus}{parameters required to fit a transition probability matrix}

\item{theta1}{parameters required to fit a DCC model}

\item{theta2}{parameters required to fit a DCC model}

\item{returns}{time series of returns data}

\item{numRegimes}{number of regimes to fit to the data}

\item{Q}{initial value of the correlation}

\item{Q.bar}{constant unconditional correlation matrix}
}
\description{
Calculate the log-liklihood using Expectation Maximization
}
\details{
This is an internal function. Given a set of paramters it returns the 
log-liklihood using Expectation Maximization procedure described in (Lee, 2010).
It also returns the final covariance matrix and final probability of being
in each state.
}
