% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCoux.R
\name{.huber}
\alias{.huber}
\title{Huber function}
\usage{
.huber(x, p)
}
\arguments{
\item{x}{the value}

\item{p}{number of columns}
}
\description{
Huber function
}
\details{
Calculate the output of a huber function given the value and the number of 
variables
}
