% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentSwitchingDCC.R
\name{.initFilterProb}
\alias{.initFilterProb}
\title{Calculates the unconditional state probabilities}
\usage{
.initFilterProb(P)
}
\arguments{
\item{P}{esitmate of the transition probability matrix}
}
\description{
Calculates the unconditional state probabilities
}
\details{
This is an internal function for calculating the intial steady state
probabilities of being in a state.
}
