% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentSwitchingDCC.R
\name{.initQ}
\alias{.initQ}
\title{Calculate the initial value of the correlation}
\usage{
.initQ(R, numRegimes, w)
}
\arguments{
\item{R}{xts object of returns data}

\item{numRegimes}{number of regimes to fit to the data}

\item{w}{weight vector to select entries from returns}
}
\value{
Q0 is a list of correlation estimates for each regime
}
\description{
Calculate the initial value of the correlation
}
\details{
This is an internal function for calculating the intial estimate of 
correlation. The entire data set is partioned into parts based on the 
weight vector and correlation is computed for each part.
}
