% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentSwitchingDCC.R
\name{.isdcc.loglik}
\alias{.isdcc.loglik}
\title{Calculate the log-liklihood}
\usage{
.isdcc.loglik(params, returns, Sigma, numRegimes, Q, Q.bar)
}
\arguments{
\item{params}{contains paramters for estimating the transition probability 
matrix and parameters required for fitting a DCC model to each
regime.}

\item{returns}{xts obect of returns}

\item{Sigma}{fitted volatility for each asset return}

\item{numRegimes}{number of regimes to fit to the data}

\item{Q}{initial value of the correlation}

\item{Q.bar}{constant unconditional correlation matrix}
}
\description{
Calculate the log-liklihood
}
\details{
This is an internal function that is used as an objective by the DEoptim
optimizer. Given a set of paramters it returns the negative of the 
log-liklihood.
}
