% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.mp.obj}
\alias{.mp.obj}
\title{Fitting median to the data}
\usage{
.mp.obj(theta, lambdas, gamma)
}
\arguments{
\item{theta}{parameter to be optimized. In this case it is 
gamma, (variables/observations)}

\item{lambdas}{eigenvalues to which the distribution is fitted}
}
\description{
Fitting median to the data
}
\details{
This method calculates the median of the eigenvalues given that they follow 
a Marchenko Pastur distribution. The median is matched to the 50 percentile
quantile from the MP distribution. This helps fit scale.factor to the data
}
