% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCoux.R
\name{.obj.helper}
\alias{.obj.helper}
\title{Internal helper function for the Objective}
\usage{
.obj.helper(
  smoothing.matrix,
  R,
  y.hat,
  Sigma.hat,
  startup_period,
  training_period,
  lambda
)
}
\description{
Internal helper function for the Objective
}
