% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCoux.R
\name{.obj}
\alias{.obj}
\title{Objective to calculate the optimal smoothing matrix}
\usage{
.obj(params, R, y.hat, Sigma.hat, startup_period, training_period, lambda)
}
\arguments{
\item{params}{d(d+1)/2 paramaters to be optimzed}

\item{R}{data}

\item{y.hat}{fitted yalues for the data}

\item{Sigma.hat}{value of the covarianve matrix}

\item{startup_period}{length of samples required to calculate initial values}

\item{training_period}{length of samples required to calculate forecast errors
for evalualating the objective}

\item{lambda}{known constant as described in the paper}
}
\description{
Objective to calculate the optimal smoothing matrix
}
\details{
The objective calculates the determinant of the one step ahead forecast errors.
The objective is a very noisy function and hence the optimization is time 
consuming and replicating paramaters across runs is difficult
}
