% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentSwitchingDCC.R
\name{.optim.params}
\alias{.optim.params}
\title{Initialize control paramters}
\usage{
.optim.params(numRegimes, tau = c(2, 10), theta = c(0, 1), ...)
}
\arguments{
\item{numRegimes}{number of regimes to fit to the data}

\item{tau}{lower and upper bounds for the tau paramter required to compute the
transition probability matrix}

\item{theta}{lower and upper bounds for the theta paramter required to compute the
DCC model for each regime#'}

\item{...}{additional control paramters required by DEoptim. Several control
paramters are defualted if not passed}
}
\description{
Initialize control paramters
}
\details{
This is an internal function for initializing differnt control paramters for
the optimizer and setting up bounds for the paramters that need to be 
optimized
}
