% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCoux.R
\name{.orthogonal.matrix}
\alias{.orthogonal.matrix}
\title{Product of Givens rotoation matrices}
\usage{
.orthogonal.matrix(d, angles)
}
\arguments{
\item{d}{number of columns}

\item{theta}{angle in the range [-pi/2, pi/2]}
}
\description{
Product of Givens rotoation matrices
}
\details{
An orthogonal matrix of size d can be parameterized by d(d-1)/2 angles
}
