% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.plot.spikedCovariance}
\alias{.plot.spikedCovariance}
\title{internal function for Eigenvalue plot}
\usage{
.plot.spikedCovariance(x, norm = NA, statistical = FALSE)
}
\arguments{
\item{x}{model of the type spikedCovariance}

\item{norm}{Type of matrix norm that must be calculated. If missing value from
the model will be used}

\item{statistical}{Stein/Entropy/Divergence/Affinity/Frechet. Default is set to NA.
when a valid value is set norm and pivot values are ignored}
}
\description{
internal function for Eigenvalue plot
}
\details{
Compares shrunk eigenvalues against sample eigenvalues
}
