% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.s.ell}
\alias{.s.ell}
\title{Equation 6.3}
\usage{
.s.ell(lambda, gamma)
}
\arguments{
\item{lambda}{sample eigenvalue}

\item{gamma}{fitted value of variables/observations}
}
\description{
Equation 6.3
}
\details{
Equation 6.3 as described in (Donoho, Gavish, and Johnstone, 2013)
}
