% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.shrink.eigen}
\alias{.shrink.eigen}
\title{Table 2}
\usage{
.shrink.eigen(lambda, gamma, type)
}
\arguments{
\item{lambda}{sample eigenvalue}

\item{gamma}{fitted value of variables/observations}

\item{type}{combination of (pivot+norm)/(statistical loss)}
}
\description{
Table 2
}
\details{
Table 2 as described in (Donoho, Gavish, and Johnstone, 2013)
}
