% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.shrink.eigen2}
\alias{.shrink.eigen2}
\title{Optimization problem 6.2}
\usage{
.shrink.eigen2(ell, c, s, type, lambda)
}
\arguments{
\item{type}{combination of (pivot+norm)/(statistical loss)}

\item{lambda}{sample eigenvalue to be shrunk}

\item{gamma}{fitted value of variables/observations}
}
\description{
Optimization problem 6.2
}
\details{
Some of the shrinkers do not have an analytical solution and the non-lineartiy
must be computed numerically. This function implements
Optimization problem 6.2 as described in (Donoho, Gavish, and Johnstone, 2013)
}
