% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{.sigma.sq.est}
\alias{.sigma.sq.est}
\title{Calculate unknown noise estimate}
\usage{
.sigma.sq.est(lambdas, numObs, numOfSpikes)
}
\arguments{
\item{lambdas}{eigenvalues to which the distribution is fitted}

\item{numObs}{number of data observations}

\item{numOfSpikes}{number of spikes in the spike covariance model}
}
\description{
Calculate unknown noise estimate
}
\details{
This method calculates the unknown unbiased estimate of noise for the eigenvalues
based on the procedure described in (Kritchman and Nadler, 2009). It involves
solving a couple of nonlinear equations.
}
