% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCoux.R
\name{.smoothing.matrix}
\alias{.smoothing.matrix}
\title{Smoothing matrix}
\usage{
.smoothing.matrix(params, d)
}
\arguments{
\item{params}{a vector of d(d+1)/2 parameters. First d are the eigenvalues and
remaining d(d-1)/2 are the angles.}

\item{d}{number of columns}
}
\description{
Smoothing matrix
}
\details{
The smooting matrix can be parametrized using the d eigenvalues and d(d-1)/2
angles to represent the orthogonal matrix when carrying out a spectral
decomposition
}
