% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stambaugh.R
\name{.stambaugh.dist}
\alias{.stambaugh.dist}
\title{An internal function that is used to calculate the Mahalanobis distance}
\usage{
.stambaugh.dist(data, model, level, id.n = 10)
}
\arguments{
\item{data}{data used to fit the covariance}

\item{model}{fitted models for covariance}

\item{id.n}{number of outliers to show}

\item{confidence}{level for the test}
}
\description{
An internal function that is used to calculate the Mahalanobis distance
}
\details{
The function takes in the model, data and the significance level and calculated
the critical values and the Mahalanobis distance.
}
\author{
Rohit Arora
}
