% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/independentSwitchingDCC.R
\name{.TransitionProb}
\alias{.TransitionProb}
\title{Create a transition probability matrix for the states in the Hidden Markov Model}
\usage{
.TransitionProb(tau, numRegimes)
}
\arguments{
\item{tau}{parameter that is optimized}

\item{numRegimes}{number of regimes to fit to the data}
}
\description{
Create a transition probability matrix for the states in the Hidden Markov Model
}
\details{
Each row in the transition probability matrix sums to 1. For N regimes
we create N(N-1) paramters.
}
