% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{dow30data}
\alias{dow30data}
\title{Symbol Data}
\source{
Yahoo finance
}
\usage{
dow30data
}
\description{
Contains returns for 30 symbols found in Dow Jones 30
             starting from 2014-04-01 to 2015-07-23. Data maybe updated
             using the getSymbolData script in the inst folder
}

