% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{estSpikedCovariance}
\alias{estSpikedCovariance}
\title{(Donoho, Gavish, and Johnstone, 2013)}
\usage{
estSpikedCovariance(R, gamma = NA, numOfSpikes = 1, norm = c("Frobenius",
  "Operator", "Nuclear"), pivot = 1, statistical = NA,
  standardize = FALSE)
}
\arguments{
\item{R}{xts object of asset returns}

\item{gamma}{ratio of varibales/observations. If NA it will be estimated from
the data. One can set it to variables/observations}

\item{numOfSpikes}{number of spikes in the spike covariance model.
It is defaulted to 1}

\item{norm}{Type of matrix norm that must be calculated. Defaults to Frobenius}

\item{pivot}{takes values from 1...7. Details can be found in the paper}

\item{statistical}{Stein/Entropy/Divergence/Affinity/Frechet. Default is set to NA.
when a valid value is set norm and pivot values are ignored}

\item{standardize}{If true eigenvalues will be scaled such that the scale
                   factor becomes 1}
}
\description{
(Donoho, Gavish, and Johnstone, 2013)
}
\author{
Rohit Arora
}

