% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{etfdata}
\alias{etfdata}
\title{Symbol Data}
\source{
Yahoo finance
}
\usage{
etfdata
}
\description{
Contains returns for 9 etf symbols XLE, XLY, XLP, XLF, XLV,XLI,
             XLB, XLK,XLU starting from 2008-01-01 to 2010-12-31. Data maybe
             updated using the getSymbolData script in the inst folder
}

