% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{factor.data}
\alias{factor.data}
\title{Factor Data}
\source{
Ken.French Website, Lubos Pastor website, Yahoo finance
}
\usage{
factor.data
}
\description{
Contains 8 factors. 5 Fama-French factors. 1 Momentum factor.
             1 Liquidity factor, 1 Volatility factor.
}

