% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fmmc.R
\name{fmmc.cov}
\alias{fmmc.cov}
\title{This is an implementation of covariance matrix estimation using Factor Model
Monte Carlo method as described in Jiang and Martin (2013).}
\usage{
fmmc.cov(R, factors, robust = FALSE, parallel = TRUE, align = c("end",
  "begin"), ...)
}
\arguments{
\item{R}{vector of asset returns in xts format}

\item{factors}{matrix of factor returns in xts format}

\item{robust}{boolean to indicate if robust methods must be used for fitting
factor models and constucting the covariance matrix. By default this
option is turned off.}

\item{parallel}{boolean to indicate if all cores on the system must be used}

\item{align}{string to indicate where the longer backfilled return histories
must be truncated when available returns are unequal. The default value
is to align them at the end
 truncated to align the returns}

\item{...}{allows passing paramters to factorAnalytics to butild the factor
model or control parameters to covrob for constructing the robust
covariance matrix}
}
\description{
This is an implementation of covariance matrix estimation using Factor Model
Monte Carlo method as described in Jiang and Martin (2013).
}
\details{
This method takes in returns data and factors as time series. The returns
and factor time-series must be aligned. The method is applicable
when we have longer factor histories and shorter asset return histories. In
this case past returns can be backfilled using the FMMC methodology. We can
apply FMMC to all the assets in the portfolio. Once the returns are backfilled
we can construct a more accurate measure of classical and robust covariance
using backfilled returns. Backfilled returns for different assets must be
aligned by truncating date in the begining or end is returns have unequal
available return histories.
}
\author{
Rohit Arora
}

