% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{largesymdata}
\alias{largesymdata}
\title{Symbol Data}
\source{
Yahoo finance
}
\usage{
largesymdata
}
\description{
Contains returns for 79 random symbols found in S&P 500
             starting from 2007-04-01 to 2015-07-02. Data maybe updated
             using the getSymbolData script in the inst folder
}

