% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LWLinearShrinkgae.R
\name{lw_linear_shrinkage}
\alias{lw_linear_shrinkage}
\title{Calculate the shrinkage estimator of covariance using the procedure
described in (Ledoit, Olivier, and Michael Wolf. "Honey, I shrunk
the sample covariance matrix.".}
\usage{
lw_linear_shrinkage(X, demean = TRUE, robust = FALSE)
}
\arguments{
\item{X}{Data matrix.}

\item{demean}{remove mean from the features}

\item{robust}{use robust estimator of average correlation}
}
\value{
A list containing the shrinkage estimator of the covariance matrix
        and the shrinkage intensity.
}
\description{
Calculate the shrinkage estimator of covariance using the procedure
described in (Ledoit, Olivier, and Michael Wolf. "Honey, I shrunk
the sample covariance matrix.".
}
\details{
This function calculates the Ledoit-Wolf shrinkage covariance estimator
to constant correlation unequal variance target matrix.
}
\author{
Rohit Arora
}
