% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/independentSwitchingDCC.R
\name{plot.isdcc}
\alias{plot.isdcc}
\title{Implied State plot}
\usage{
\method{plot}{isdcc}(x, y = c(1, 2), ...)
}
\arguments{
\item{x}{model of the type isdcc obtained by fitting an IS-DCC model to the data}

\item{y}{type of plot. takes values 1/2. 1 = Implied States, 2 = Smoothed Proabability}

\item{...}{additional arguments unused}
}
\description{
Implied State plot
}
\details{
Plot implied states using the fitted Independent Switching DCC model
}
\examples{
\dontrun{
 data("largereturn")
 model <- isdccfit(largesymdata, numRegimes=2)
 plot(model)
}
}
\author{
Rohit Arora
}

