% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stambaugh.R
\name{plot.missing}
\alias{plot.missing}
\title{Plot data to visualize missing values}
\usage{
\method{plot}{missing}(data, which = c(3, 4))
}
\arguments{
\item{data}{a matrix, data-frame or an xts/zoo object}

\item{which}{takes values 3/4. 3 = Summary plot, 4 = Matrix plot}
}
\description{
Plot data to visualize missing values
}
\details{
This method takes in data as a matrix,data frame or an xts object and
plots summary or matrix plots for data. Missing values highlighted in red.
}
\author{
Rohit Arora
}

