% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stambaugh.R
\name{plot}
\alias{plot}
\title{Generic plot method for Stambaugh type objects}
\usage{
plot(x, which.plots = c(1, 2), ...)
}
\arguments{
\item{x}{Stambaugh objects}

\item{...}{}
}
\description{
Generic plot method for Stambaugh type objects
}

