% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/randomMatrix.R
\name{plot.RMT}
\alias{plot.RMT}
\title{Eigenvalue plot}
\usage{
\method{plot}{RMT}(x, y, ...)
}
\arguments{
\item{x}{model of the type RMT obtained by fitting an RMT model to the data}

\item{y}{unused}

\item{...}{additional arguments unused}
}
\description{
Eigenvalue plot
}
\details{
Plots eigenvalues of the correlation matrix and overlays the Marchenko-Pastur
density on top of it. There is a shap cutoff for the density. We are concerned
with eigenvalues beyond this cutoff. Paramters used for plotting are added
to the plot
}
\examples{
\dontrun{
 data("largereturn")
 model <- estRMT(largesymdata)
 plot(model)
}
}
\author{
Rohit Arora
}

