% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{plot.spikedCovariance}
\alias{plot.spikedCovariance}
\title{Eigenvalue plot}
\usage{
plot.spikedCovariance(R, gamma = NA, numOfSpikes = 1, ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{gamma}{atio of varibales/observations. If NA it will be estimated from
the data. One can set it to variables/observations}

\item{...}{additional arguments unused}

\item{numberOfSpikes}{model of the type spikedCovariance}
}
\description{
Eigenvalue plot
}
\details{
Compares shrunk eigenvalues against sample eigenvalues for all norms and losses
}
\examples{
\dontrun{
 data("rmtdata")
 model <- estSpikedCovariance(rmtdata, numOfSpikes=10)
 plot(model)
}
}
\author{
Rohit Arora
}

