% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stambaugh.R
\name{plot.stambaugh}
\alias{plot.stambaugh}
\title{Plot Ellipsis or Distance plot for the Stambaugh estimator}
\usage{
\method{plot}{stambaugh}(x, y = c(1, 2), ...)
}
\arguments{
\item{x}{fitted models for covariance}

\item{y}{takes values 1/2. 1 = Ellipse plot, 2 = distance plot}

\item{...}{allows passing additional paramters}
}
\description{
Plot Ellipsis or Distance plot for the Stambaugh estimator
}
\details{
This method takes in fitted models and a paramter for deciding the type of plot
}
\author{
Rohit Arora
}

