% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stambaugh.R
\name{plotmissing}
\alias{plotmissing}
\title{Plot data to visualize missing values}
\usage{
plotmissing(data, which = c(3, 4))
}
\arguments{
\item{data}{an xts/zoo object}

\item{which}{takes values 3/4. 3 = Summary plot, 4 = Matrix plot}
}
\description{
Plot data to visualize missing values
}
\details{
This method takes in data as an xts object and plots the data.
Missing values highlighted in red for matrix plot and time series of returns
are shown in in Summary plot
}
\author{
Rohit Arora
}

