% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spikedCovariance.R
\name{plotSpikedCovariance}
\alias{plotSpikedCovariance}
\title{Eigenvalue plot. Similar to figure 1 in the paper}
\usage{
plotSpikedCovariance(R, gamma = NA, numOfSpikes = NA, method = "KNTest",
  ...)
}
\arguments{
\item{R}{xts object of asset returns}

\item{gamma}{ratio of varibales/observations. If NA it will be set to
ratio of varibales/observations}

\item{numOfSpikes}{model of the type spikedCovariance}

\item{method}{KNTest/median-fitting. Default is KNTest}

\item{...}{additional arguments unused}
}
\description{
Eigenvalue plot. Similar to figure 1 in the paper
}
\details{
Compares shrunk eigenvalues against sample eigenvalues for all norms and losses
}
\examples{
\dontrun{
 data("rmtdata")
 plot(rmtdata, numOfSpikes=10)
}
}
\author{
Rohit Arora
}

