% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rmtdata}
\alias{rmtdata}
\title{Simulated data for Spiked Covarianve Model}
\usage{
rmtdata
}
\description{
Contains large data for 500 observations and 100 variables based
               on standard normal independent random variables
}

