% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stambaugh.R
\name{stambaugh.ellipse.plot}
\alias{stambaugh.ellipse.plot}
\title{Plot Ellipsis for the Stambaugh estimator}
\usage{
stambaugh.ellipse.plot(models)
}
\arguments{
\item{models}{fitted models for covariance}
}
\description{
Plot Ellipsis for the Stambaugh estimator
}
\details{
This method takes in fitted models for Stamgaugh Estimator. It then plots a
comparison of the fitted models using ellipsis plot
}
\author{
Rohit Arora
}

