% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stambaugh.R
\name{stambaugh.est}
\alias{stambaugh.est}
\title{Implement Stamgaugh Covariance estimate for multiple starting dates}
\usage{
stambaugh.est(R, ...)
}
\arguments{
\item{R}{xts or matrix of asset returns}
}
\description{
Implement Stamgaugh Covariance estimate for multiple starting dates
}
\details{
This method takes in data as a matrix or an xts object where multiple
time series with different starting dates are merged together. It then
computes a covariance estimator as described in Stambugh (1997). Covariance
estimate can also be robustified
}
\author{
Rohit Arora
}

