% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stambaugh.R
\name{stambaugh.fit}
\alias{stambaugh.fit}
\title{Estimate covariance matrices using Stambaugh method for classical and Robust
methods}
\usage{
stambaugh.fit(R, method = c("classic", "robust", "truncated"), ...)
}
\arguments{
\item{R}{xts or matrix of asset returns}

\item{method}{type of model to fit. Takes 3 values classic/robust/truncated}

\item{...}{pass paramters to fitTimeSeriesFactorModel(factorAnalytics),
covRob, lmRob (Robust) functions}
}
\description{
Estimate covariance matrices using Stambaugh method for classical and Robust
methods
}
\details{
This method takes in data as a matrix or an xts object where multiple
time series with different starting dates are merged together. It then
computes a covariance estimator based on the specifed style
}
\author{
Rohit Arora
}

