% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{symdata}
\alias{symdata}
\title{Symbol Data}
\source{
Yahoo finance
}
\usage{
symdata
}
\description{
Contains returns for 7 symbols ie BABA, TWTR, LNKD, YHOO, GE,
             LAZ, V starting from 2007-04-01 to 2015-07-02. Data maybe updated
             using the getSymbolData script in the inst folder
}

