% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RobustCroux.R
\name{robustMultExpSmoothing}
\alias{robustMultExpSmoothing}
\title{Robust Multivariate Exponential Smoothing}
\usage{
robustMultExpSmoothing(
  R,
  corr = FALSE,
  smoothMat = NA,
  startup_period = 10,
  training_period = 60,
  seed = 9999,
  trials = 50,
  method = "L-BFGS-B",
  lambda = 0.2
)
}
\arguments{
\item{R}{An xts object containing the data.}

\item{corr}{A boolean indicating whether additional correlation matrix is needed. If true,
additional correlation matrix is returned.}

\item{smoothMat}{Optional. The optimal smoothing matrix. If missing, it is estimated. 
The procedure may be very slow for high-dimensional data. Also,
the objective function being very noisy, optimization across
multiple runs may lead to different smoothing matrices.}

\item{startup_period}{Integer. Length of samples required to calculate initial values.}

\item{training_period}{Integer. Length of samples required to calculate forecast errors
for evaluating the objective if the smoothing matrix is estimated.}

\item{seed}{Integer. Random seed to replicate the starting values for optimization.}

\item{trials}{Integer. Number of starting values to try for any optimization. 
A large number of trials for high dimensions can be time-consuming.}

\item{method}{Character. Optimization method to use to evaluate an estimate of 
the smoothing matrix. Default is "L-BFGS-B".}

\item{lambda}{Numeric. Known constant as described in the paper. Defaults to 0.2.}
}
\value{
A list containing:
\item{smoothMat}{The optimal smoothing matrix.}
\item{smoothValues}{The smoothed values as an xts object.}
\item{cleanValues}{The cleaned values as an xts object.}
\item{covMatList}{A list of covariance or correlation matrices for each time period.}
}
\description{
Calculate a robust estimate of the covariance matrix while also smoothing and 
cleaning the data using the procedure described in 
(Croux, Gelper, and Mahieu, 2010).
}
\author{
Rohit Arora
}
